2023-08 WIP

+++ scala2/src/main/scala/jurisk/adventofcode/y2022/Advent24.scala
@@ -34,7 +34,7 @@ object Advent24 {
         }
     }
 
-    private val cycleTime: Int = lcm(area.width - 2, area.height - 2).toInt
+    private val cycleTime: Int = lcm(area.width - 2, area.height - 2)
 
     def nextTimeWrapping(time: Int): Int = (time + 1) % cycleTime
